Add files via upload
controller including friction FF

During experimentation the "friction constant" and "error at which friction is neglected" constants can be adjusted for the best step response. 
The friction constant is the extra amount of power going to the motor. During testing the effect of the feedforward against friction can be measured.